### 2024-01-09 #6197 @editorialbot
[PRE REVIEW]: RAMP: stochastic simulation of user-driven energy demand time series
**Submitting author:** @FLomb (<a href="http://orcid.org/0000-0002-7624-5886">Francesco Lombardi</a>)
**Repository:** https://github.com/RAMP-project/RAMP
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.5.0
