### 2024-02-02 898be3c
Implemented fix for duty cylce repetition issue 78 (https://github.com/RAMP-project/RAMP/issues/78) - in Appliance -> .update_daily_use:   - duration of switch-on event is corrected to be only as long as duty cylce   - corrected duration of switch-on event is returned - in Appliance -> .generate_load_profile   - calculation of total time of operation is performed using corrected duration of switch-on event
