### 2024-02-02 df3c22f
Implemented unit test for fix of issue 78 (https://github.com/RAMP-project/RAMP/issues/78)
