### 2024-02-03 #6197 @AdamRJensen
@kyleniemeyer I'm a new JOSS editor and could take on the editorial role for this paper
