### 2024-02-07 #6197 @AdamRJensen
:wave: @willu47, @fraukewiese, & @LSRathore, would any of you be willing to review this submission for JOSS concerning energy
demand time series? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
