### 2024-02-08 #6197 @willu47
Hi @AdamRJensen - can you clarify the conflict of interest policy for JOSS. I work/have worked in a European funded project with @FLomb; although our work has nothing to do with RAMP and is in different work packages in the project. I believe I could objectively review the RAMP codebase, but this should be first declared openly.
