### 2024-02-12 #6197 @AdamRJensen
@willu47 You can find the JOSS guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) and it sounds like your case could be perceived as being borderline though I don't doubt your abilities to be objective. I'll attempt to find another reviewer, but thank you very much for your interest 😄

@noah80 could you be interested in reviewing?
