### 2024-02-27 #6197 @AdamRJensen
👋 @fabianhofmann and @trevorb1 I see that you both have published with JOSS and was wondering if you be willing to review this submission concerning energy demand time series? This is a great way to give back to JOSS 😄

As you know, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
