### 2024-02-27 #6418 @editorialbot
[REVIEW]: RAMP: stochastic simulation of user-driven energy demand time series
**Submitting author:** @FLomb (<a href="http://orcid.org/0000-0002-7624-5886">Francesco Lombardi</a>)
**Repository:** https://github.com/RAMP-project/RAMP
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.5.2
