### 2024-02-29 76e5342
Fix issue 114. Skip app if func_time or rand_time==0
