### 2024-02-29 1f87cc0
Add warning if appliance with func_time=0 is defined
