### 2024-02-29 d9fcb0a
Add tests and fix faulty func_time warning
