### 2024-03-05 #6418 @FLomb
> @FLomb: Could you clarify the individual contributions of the authors?

Hi @noah80, the contributions are the following (thanks for asking; we didn't know they were requested):

- F. Lombardi: conceptualisation, code development, code maintenance, code review, documentation, project coordination
- PF. Duc: code development, code maintenance, code review, documentation, project coordination
- M.A. Tahavori: code development, code maintenance, code review, documentation
- C. Sanchez-Solis: code development, code testing, code review
- S. Eckhoff: code development (interface), code testing
- M.C.G. Hart: code development (interface), code testing
- F. Sanvito: conceptualisation (mobility), code development (mobility), code testing, project coordination
- G. Ireland: code testing, code review, feature coordination
- S. Balderrama: code testing, code review, project coordination
- S. Quoilin: conceptualisation, code review
