+++ ramp/core/core.py
@@ -1029,7 +1029,6 @@ class Appliance:
             2. power array size is not (366,1)
         """
 
-
         self.user = user
         self.name = name
         self.number = number
@@ -1037,7 +1036,9 @@ class Appliance:
 
         if func_time == 0:
             warnings.warn(
-                UserWarning(f"Func_time of appliance '{self.name}' is defined as 0. Ignore if this is intended")
+                UserWarning(
+                    f"Func_time of appliance '{self.name}' is defined as 0. Ignore if this is intended"
+                )
             )
 
         self.func_time = func_time
@@ -1854,7 +1855,9 @@ class Appliance:
         ]
 
         tot_time = 0
-        while tot_time <= rand_time and rand_time != 0:  # Check if rand_time is 0 (-> fix issue 114)
+        while (
+            tot_time <= rand_time and rand_time != 0
+        ):  # Check if rand_time is 0 (-> fix issue 114)
             # one option could be to generate a lot of them at once
             indexes = self.rand_switch_on_window(
                 rand_time=rand_time,  # TODO maybe only consider rand_time-tot_time ...
