+++ tests/test_func_time_zero.py
@@ -2,6 +2,7 @@ import numpy as np
 import pytest
 from ramp import UseCase, User
 
+
 def test_skip_when_func_time_zero():
 
     # Generate test user and appliance
@@ -14,22 +15,22 @@ def test_skip_when_func_time_zero():
     )
 
     # Add to use_case
-    use_case = UseCase(
-        name='test_use_case',
-        users=[user]
-    )
+    use_case = UseCase(name="test_use_case", users=[user])
 
     # Calculate peak time range for this use_case
     peak_time_range = use_case.calc_peak_time_range()
 
     # Generated one load profile of this appliance
     power = 1000
-    appliance.generate_load_profile(prof_i=0, peak_time_range=peak_time_range, day_type=0, power=power)
+    appliance.generate_load_profile(
+        prof_i=0, peak_time_range=peak_time_range, day_type=0, power=power
+    )
 
     # Check that no use of this appliance is simulated -> the appliances load profile is always smaller than it's power
     # (Checking for the load_profile to always be 0 might be unreliable, since the RAMP core "marks" potential usage
     # windows with small power values larger 0)
     assert np.max(appliance.daily_use) < power
+
 
 def test_warning_when_func_time_zero():
     user = User(user_name="Test User", num_users=1)
@@ -39,6 +40,7 @@ def test_warning_when_func_time_zero():
         func_cycle=20,
         time_fraction_random_variability=0.1,
     )
+
 
 with pytest.warns():
     test_warning_when_func_time_zero()
