### 2024-03-14 0c65148
Check for raising of warning within the test function
