### 2024-03-14 afe1d02
Move duty_cycle selection from update_daily_use to rand_switch_on_window
- Introduce current_duty_cycle_id as attribute of Appliance to
keep track of the current duty cycle and avoid re-evaluating
which duty-cycle is concerned by a switch on event
- Change x in range(a,b) to a <= x < b as the latter is about 5 times
faster 243 ns vs 51.4 ns
- Remove direct mention of specific issue in the comments
