### 2024-03-14 293242a
Compute total time before while loop
As now indexes are not updated by update daily_use but by
rand_switch_on_window, it is less risky for backward compatibility
to update the total time before entering the loop
