### 2024-03-14 1355a16
Keep assignement to duty cycle 3 similar to past code
