### 2024-03-15 #6418 @FabianHofmann
## Comment and Questions

Hey @FLomb, very nice project! To get the ball rolling here is the first round of comments from my side.

### General checks
1. Some of the authors do not appear in the contributions list of the Github repository. Could you clarify their role? https://github.com/RAMP-project/RAMP/graphs/contributors
2. JOSS asks for reproducibility of the figure in the paper (Data Sharing & Reproducibility). Is the code online somewhere? I could not see it in the `joss_paper`  branch of the repo.

### Functionality

1. The readme proposes to create a python environment with python=3.8. This python version is soon outdated. Is the software supporting newer python versions? If yes, I would suggest updating the python version in the Readme.
3. After having followed the installation instructions, which work fine, I had a look at the example at https://github.com/RAMP-project/RAMP?tab=readme-ov-file#building-a-model-with-a-python-script. What I find confusing is that I cannot display the `household_1` object. The error message is clear (there are no appliances), but for newbies this might be confusing. Also empty objects which seems to be that basic to the whole software should be displayable.

### Documentation

The Readthedocs documentation is great with all its examples. Two minor remarks.
1. there are some examples which do not have an intro at the beginning, like the fixed flat-appliance example. These are however very valuable to understand the context.
3. the navigation on the left sidebar (with the partial TOC) is a bit counter-intuitive, I find. I am getting lost easily without knowing where I am. If you agree, I would suggest increasing the scope of the sidebar's TOC.
4. What is the test coverage of the project? Since there are not that many, it would be good to check that minimum coverage is realised.
5. I cannot find a contribution guideline as required by JOSS.

### Software paper

1. The State of the field is not realised. It would be interesting what other alternative software is there and how it compares.

Again, nice collaborative project!
