+++ .github/workflows/python-pypi-test.yml
@@ -14,6 +14,9 @@ jobs:
   build_and_deploy:
     name: Build the release and deploy to test pypi
     runs-on: ubuntu-latest
+    permissions:
+      id-token: write
+      contents: read
     steps:
       - name: Checkout code
         uses: actions/checkout@master
