### 2024-03-18 2e162de
Fix the duty cycle test
