+++ .github/workflows/python-pypi-test.yml
@@ -7,7 +7,12 @@ on:
       - main
     tags: # Add tag event trigger
       - '*'
-      
+  pull_request:
+    branches:
+      - development
+      - main      
+    tags: # Add tag event trigger
+      - '*'
 jobs:
   build_and_deploy:
     name: Build the release and deploy to test pypi
