### 2024-03-18 d5029ea
Make continuous_duty_cycle an Appliance attribute
