### 2024-03-19 35428a4
Add continuous_use_duty_cycle to appliance attributes
