### 2024-03-19 5b42bdc
Prevent load from excel to throw error if new attribute not present
