### 2024-03-28 #6418 @trevorb1
# Comments

Hi @FLomb; thanks so much for submitting RAMP and suggesting me to review! The collaboration, documentation, and existing use in literature of the RAMP project is super nice! Please find below my comments.

## General Checks
1. Same question as @FabianHofmann; is there code where I can reproduce Figure 1?
2. Your reference of "Generating high-resolution multi-energy load profiles for remote areas with an open-source stochastic model" is a 2019 paper in Energy describing the first version of RAMP and validating its functionality, I believe? If so, can you please describe (high level) how this paper is different from the one mentioned? This doesn't necessarily need to go in the paper, just for my reference clarifying the differences please. (Sidenote; I know other packages like pyam are published in different journals - ORE and JOSS in that case - so Im not trying to attack this submission or anything! Just trying to understand the differences.)

## Functionality
1. In the ["Using real calendar days to generate profiles"](https://rampdemand.readthedocs.io/en/latest/examples/year_simulation/year_simulation.html) example, it walks through how to use the CLI to generate a number of profiles. I tried to use the excel file generated from the ["Using tabular inputs to build a model"](https://rampdemand.readthedocs.io/en/latest/examples/using_excel/using_excel.html) as an input, but received a matplotlib error for the second plot. The command I ran and the error I received is shown below. I can attach the full trace back if that helps. Not sure if I am just misunderstanding what my input file should be?

```bash
$ ramp -i example_excel_usecase.xlsx -n 10
ValueError: x and y must have same first dimension, but have shapes (1440,) and (1,)
```

2. I couldn't find where package dependencies are listed. When I tried to run tests, I got a `ModuleNotFoundError: No module named 'scipy'` error. The only place I saw dependences was in the `ramp/__init__.py` file, but `scipy` is not listed. If I `pip install scipy` into my environment and run `pytest tests/` I get a `ModuleNotFoundError: No module named 'nbconvert'`. I had to `pip install nbconvert` as well to get tests to run. Clarifying the dependencies for using vs. developing would help!

3. If I run tests (with the extra pip installs), I get one failure:

```bash
FAILED tests/test_switch_on.py::TestRandSwitchOnWindow::test_coincidence_normality_on_peak - AssertionError: The 'coincidence' values are not normally distributed.

assert 0.029748765751719475 > 0.05
```

## Documentation
1. In the Quick Start section, appliances are added through the method `User.Appliance(...)`. Later on in the examples section, appliances are added through the method `User.add_appliance(...)`. Based on the API reference, the `Appliance(...)` method should only be used for when working with legacy code? If users are encouraged to use the `add_appliance(...)` method, updating docs to reflect this would be good! (Same issue also appears in ["Appliances with multiple cycles"](https://rampdemand.readthedocs.io/en/latest/examples/multi_cycle/multi_cycle.html) example)

2. I find it a little confusing how the [documentation site contribution guidelines](https://rampdemand.readthedocs.io/en/latest/intro.html#contributing) and file [`CONTRIBUTING.md`](https://github.com/RAMP-project/RAMP/blob/main/CONTRIBUTING.md) in the repository don't match?

3. It may be worth noting in the contributing guidelines that developers should follow Black formatting, and contributions will be checked for it with your actions (although, I do see you have the Black badge on the README.md). Alternatively, having a pre-commit file, or specifying in the contributing guidelines how to install Black in VSCode (or similar) to autolint contributions would be good!

5. In contributions guidelines, it may be good to specify what to include in a new issue ticket and PR (or even create templates for these)? For example, for a new issue, do you want to know OS the user is running, version of RAMP they are running, ect.

6. In contribution guidelines on the doc site, you mention contributors should perform qualitative testing only. On the repository `CONTRIBUTING.md` file, it mentions to also run tests via pytest. Please clarify what tests contributors should be running.

7. At the bottom of the introduction page, there is a note that says "This project is under active development!". Can you please clarify what the active development means? Just want to ensure this does not mean we can't trust the results. (Sorry, I know this one is a little pedantic!)

8. The examples are great at walking the reader through the many different functions of RAMP! However, I think in the first example it would be beneficial to explicitly write out what all the different arguments in the `User(...)` and `add_appliance(...)` calls are doing. This can maybe be done through a top level description like in some of your other examples (as also suggested by @FabianHofmann). At first, especially for the `add_appliance(...)` method, it wasn't necessarily clear to me what the different arguments did. While I eventually did find the info I was looking for in the `Appliance.__init__` API ref, this was not the first place I intuitively thought to look for the information (which was the `add_appliance(...)` method API ref).

## Software Paper
1. I agree with @FabianHofmann; a comparison of similar software/tools is needed.
2. Expanding on your sentence of RAMP "features several degrees of customisations" to explicitly describe some of these customizations would be great I think! For example, being able to generate different loads for different uses (EVs, hot water, cooking ect..) is a valuable point for readers, I believe.
