### 2024-03-28 #6418 @Bachibouzouk
@FabianHofmann , @trevorb1  thanks for your insightful and helpful comments!  We will start addressing them in PRs from the `joss-paper` branch (https://github.com/RAMP-project/RAMP/tree/joss-paper) :)
