### 2024-03-28 #6418 @Bachibouzouk
> 3\. If I run tests (with the extra pip installs), I get one failure:

@trevorb1 This is actually a known issue, this tests fails randomly : https://github.com/RAMP-project/RAMP/issues/99 I will use this review to push myself to investigate it further. If you have good suggestion of methodology/articles about stochastic code testing, they are welcome :)
