### 2024-03-29 8f064fc
Python=3.8 to be outdated as the prefered python version
issue:

Python 3.8 is mentioned in the README as the prefered python version for
the installation. However, this python vesion is going to be outdated by
October 2024.

Current RAMP dev can be safely used by python 3.10.

Changes:

- README.rst --> updateing the installation guide
- environment.yml --> updating the python dependency for installing
  environment through yml file
 - setup.py --> updating the python_requires accordingly
