### 2024-03-29 cd3ac19
Adding depreciation warning to Appliance method in User class
to ensure the Back-compatibility with legacy code, the Appliance method
of User class is present. However the method is not further maintained
since the nex-gen version of the sofware. To warn the users to move
towards the new method, a DeprecationWarning is added.
