### 2024-04-03 RAMP-project/RAMP#127 @JW-Kraft
Reproducible results of the stochastic model
In todays dev meeting we discussed the issue of being able to exactly reproduce RAMP's results. This has multiple motivations:

1. Reviewers of the JOSS paper submission requested code to reproduce figure 1.
https://github.com/openjournals/joss-reviews/issues/6418#issuecomment-1999853280
I'm not sure if only the raw data of the plotted load profile is requested or the complete code sample to run the displayed test in RAMP and model the plotted load profile (starting with the input file).
If it's case 2, this will not be possible without setting the random seed, since the stochastic model will return different results on different runs.

2. @FraSanvit mentioned this would be helpful for testing.
3. [Setting and noting random seeds of stochastic models is good practice to allow for reproducibility](https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1003285#s7). We should give RAMP users the ability to do that. This way they can submit a complete code sample which exactly reproduces the results presented in their papers.

I will look into how to implement this in the next days. My first idea was to just add a optional `seed` argument to the `UseCase` class.
What do you think?
