### 2024-04-04 RAMP-project/RAMP#127 @Bachibouzouk
> My first idea was to just add a optional `seed` argument to the `UseCase` class.
> What do you think?

I think it would make sense. Now a `User` instance not connected to a `UseCase` instance gets a default one as soon as `generate_single_load_profile()` method is called without providing a `peak_time_range` argument.

We should maybe look what happens if an `Appliance` instance is created without being linked to a `User`. I think the user and usecase should be created, if not provided, only when a method to generate load profiles is used
