### 2024-04-04 #6418 @Bachibouzouk
> 1. In the ["Using real calendar days to generate profiles"](https://rampdemand.readthedocs.io/en/latest/examples/year_simulation/year_simulation.html) example, it walks through how to use the CLI to generate a number of profiles. I tried to use the excel file generated from the ["Using tabular inputs to build a model"](https://rampdemand.readthedocs.io/en/latest/examples/using_excel/using_excel.html) as an input, but received a matplotlib error for the second plot. The command I ran and the error I received is shown below. I can attach the full trace back if that helps. Not sure if I am just misunderstanding what my input file should be?

@trevorb1 - this is now fixed in #https://github.com/RAMP-project/RAMP/pull/126

How should we proceed, you write on the PR that you agree with the proposed changes and then we merge them?
I guess it make sense to link to those PRs at least once in this thread so that the history of the review is all accessible form the joss-review repository
