### 2024-04-04 1227ad5
Add _add_appliance_instance method to add an Appliance to a User
