### 2024-04-04 7b76c8c
Add _add_appliance_instance method to add an Appliance to a User
