### 2024-04-04 3ee275b
Fix comparison fonction of Appliance class
The comparisons were not saved so the return value was always "True"
