### 2024-04-04 b3b495a
Fix comparison fonction of Appliance class
The comparisons were not saved so the return value was always "True"
