### 2024-04-05 #6418 @FLomb
> JOSS asks for reproducibility of the figure in the paper (Data Sharing & Reproducibility). Is the code online somewhere? I could not see it in the joss_paper branch of the repo.

> Same question as @FabianHofmann; is there code where I can reproduce Figure 1?

Thank you both for pointing this out; we had somehow missed this requirement. We will address this issue shortly.

> 2\. Your reference of "Generating high-resolution multi-energy load profiles for remote areas with an open-source stochastic model" is a 2019 paper in Energy describing the first version of RAMP and validating its functionality, I believe? If so, can you please describe (high level) how this paper is different from the one mentioned? This doesn't necessarily need to go in the paper, just for my reference clarifying the differences please. (Sidenote; I know other packages like pyam are published in different journals - ORE and JOSS in that case - so Im not trying to attack this submission or anything! Just trying to understand the differences.)

Sure, happy to provide further information about that here. In the 2019 paper, we introduced the model from a 'scientific' perspective (i.e., what are the rationale, the math, and the algorithms behind it), discussed its merits over existing alternative demand simulation approaches, and validated it against metered data. Nonetheless, from the software point of view, the code back then was just a single Python file that, despite being made openly available on GitHub, did not meet any of the standards of reproducible scientific software, nor was intended for use beyond the scope of the kind of application proposed in the paper. In other words, in the original paper, the concept and the 'science' behind the model were peer-reviewed, but the software itself was not and was not meant to.

Over the years, and thanks to the collaboration of more institutions who joined the project, RAMP has grown into a more professional, documented and tested software, including more functionalities and opportunities for customisation and for broad application. Having achieved this degree of software quality, with this paper, we set out to let the software itself undergo a peer review process. In doing so, we follow in the footsteps of many other projects that were initially published as supporting code for scientific publications in a rough, limitedly documented form and only later developed into more professional software published in JOSS (for example, Calliope: https://doi.org/10.21105/joss.00825).
