### 2024-04-07 #6418 @trevorb1
> How should we proceed, you write on the PR that you agree with the proposed changes and then we merge them?

@Bachibouzouk, sounds great! Your fix solved the problem I described. However, I ran into another issue completing the rest of the CLI examples. I left a comment in https://github.com/RAMP-project/RAMP/pull/126 describing the issue. Thanks!

> Sure, happy to provide further information about that here.

Thank you so much for the detailed description, @FLomb! This fully answers my question, as I am now clear on the differences behind the Energy and JOSS publications of RAMP. It is really great to see so much collaboration behind this project!
