### 2024-04-08 0e6e58a
Fix multi-year in cli
issue: the user could not simulate more than one year because of a
requirement that the number of power values for an appliance should not
exceed 366

solution: the checks were moved to a function of Appliance class which
is called when the usecase is initialized. If the Appliance instance has
constant power, its power timeseries is adjusted and a warning is
issued. However if the power timeseries was provided by user and does
not contain enough values to cover the simulation range, an error is
raised.
