### 2024-04-08 7703d0d
Add test for multi-year input in cli
