### 2024-04-08 5ce8fb1
Fix seasonality input of RAMP cli
Also provide a test of the CLI with example file
