### 2024-04-10 9ccd839
Adapt Testing guidelines
Issue: it was not easy to find that there were extra dependencies for
testing

Solution: mention it explicitely in the CONTRIBUTING.md
