+++ CONTRIBUTING.md
@@ -44,17 +44,31 @@ Please try to write clear commit messages. One-line messages are fine for small 
 
 ## Testing
 
-We have a qualitative testing functionality that allows to compare the results arising from a modified version of the code against default ones, for the 3 reference input files provided within the code itself.
+Testing is used by RAMP developers to make sure their new feature/bug fix is not breaking existing code. As RAMP is stochastic some tests are only qualitative, other unit tests are ran by GitHub Actions. 
 
-This functionality is accessible via `test/test_run.py`. To run the qualitative test, you'll have to go through the following steps:
-   1. run your modified code for the 3 reference input files for 30 days each. This will create 3 corresponding output files in the `results` folder
-   2. run `test/test_run.py` and visualise the comparison between the results of your code and those obtainable with the latest stable version
+Before running the tests locally, you need to install the testing dependencies
+
+```
+pip install -r tests/requirements.txt
+```
+
+### Qualitative testing
+
+The qualitative testing functionality allows to compare the results arising from a modified version of the code against default ones, for the 3 reference input files provided within the code itself.
+
+To run the qualitative test, you'll have to run
+ ```
+ python ramp/test/test_run.py
+ ```
+from the root level of this repository.
+
+If you already ran this script, you will be asked if you want to overwrite the results files (if you decide not to, the results are not going to be regenerated from your latest code version). You should compare the results of your code and those saved from the latest stable version thanks to the image which is displayed after the script ran.
    
 Ideally, the difference between reference and new results should be minimal and just due to the stochastic nature of the code. If more pronounced, it should be fully explainable based on the changes made to the code and aligned to the expectations of the developers (i.e. it should reflect a change in the output *wanted* and precisely *sought* with the commit in question).
 
 ### Unit tests
 
-Install `pytest` (`pip install pytest`) and run `pytest tests/` form the root of the repository to run the unit tests
+Run `pytest tests/` form the root of the repository to run the unit tests.
 
 ## Attribution
 
