### 2024-04-10 9d28c32
Correct the sampling of coincident switch-on on peak
Related to https://github.com/RAMP-project/RAMP/issues/99
