### 2024-04-10 4559bcf
Correct the sampling of coincident switch-on on peak
Related to https://github.com/RAMP-project/RAMP/issues/99
