### 2024-04-10 #6418 @Bachibouzouk
> @trevorb1 This is actually a known issue, this tests fails randomly : [RAMP-project/RAMP#99](https://github.com/RAMP-project/RAMP/issues/99) I will use this review to push myself to investigate it further. If you have good suggestion of methodology/articles about stochastic code testing, they are welcome :)

@trevorb1 - this should now be addressed in https://github.com/RAMP-project/RAMP/pull/129, specifically on https://github.com/RAMP-project/RAMP/pull/129/commits/b4e728a4d82804a1009b0dde4dadabae9246ae77
