### 2024-04-11 b1114a9
Improved fixed_cycle docs
