### 2024-04-12 0d72c6f
Fix small typo in core.py
Remove typo in a string
