### 2024-04-12 c1552ec
Fix small typo in core.py
Remove typo in a string
