### 2024-04-29 #6418 @Bachibouzouk
> 1. The readme proposes to create a python environment with python=3.8. This python version is soon outdated. Is the software supporting newer python versions? If yes, I would suggest updating the python version in the Readme.

@FabianHofmann - https://github.com/RAMP-project/RAMP/pull/123 bumps the python version used in RAMP to 3.10
