### 2024-04-29 #6418 @Bachibouzouk
> 5\. In contribution guidelines on the doc site, you mention contributors should perform qualitative testing only. On the repository `CONTRIBUTING.md` file, it mentions to also run tests via pytest. Please clarify what tests contributors should be running.

@trevorb1  - This has been addressed in https://github.com/RAMP-project/RAMP/pull/129 more specifically in https://github.com/RAMP-project/RAMP/pull/129/commits/9ccd839e6f0edfe7bb00d14a51301ebf4fcbc439
