### 2024-04-29 #6418 @Bachibouzouk
> 2\. After having followed the installation instructions, which work fine, I had a look at the example at https://github.com/RAMP-project/RAMP?tab=readme-ov-file#building-a-model-with-a-python-script. What I find confusing is that I cannot display the `household_1` object. The error message is clear (there are no appliances), but for newbies this might be confusing. Also empty objects which seems to be that basic to the whole software should be displayable.

@FabianHofmann - This issue should be addressed within https://github.com/RAMP-project/RAMP/pull/122
