+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,17 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: "[Bug] Bug_short_description"
+labels: bug
+assignees: ''
+
+---
+
+Describe_the_bug
+
+Checklist to make sure that the bug report ist complete:
+- [ ] OS: *your_operating_system*, *your_distribution*
+- [ ] RAMP version or Branch: *RAMP version* or *branch_name*, updated on *update_date*
+- [ ] If applicable: Attach full error message
+- [ ] If applicable: Share screenshots/images of your problem
+- [ ] If applicable: Share used input data
