### 2024-04-29 cec525b
Update workflow upon PR opening
