### 2024-05-01 4096f8f
improving the readthedocs template
within the current template, 'basicstrap', the navigation on the left sidebar (with the partial TOC) is a bit
counter-intuitive. The new template provides a more intituitive
sidebars.
