+++ docs/source/conf.py
@@ -43,7 +43,10 @@ extensions = [
     "sphinx.ext.doctest",
     "sphinx_copybutton",
     "sphinx.ext.autosectionlabel",
+    "sphinx_wagtail_theme"
 ]
+
+
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
@@ -60,7 +63,8 @@ exclude_patterns = []
 # a list of builtin themes.
 #
 extensions.append("sphinxjp.themes.basicstrap")
-html_theme = "basicstrap"
+
+html_theme = 'sphinx_wagtail_theme'
 # html_theme_path = [sphinx_pdj_theme.get_html_theme_path()]
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
