+++ docs/docs-requirements.txt
@@ -11,9 +11,7 @@ sphinx >= 1.6.4
 nbsphinx
 ipykernel
 sphinx-autobuild
-sphinx-pdj-theme # TODO to comment
 sphinx-copybutton
-sphinxjp.themes.sphinxjp
-sphinxjp.themes.basicstrap
+sphinx_wagtail_theme
 nbformat
 nbconvert
