### 2024-05-01 1cd3e7a
fixing extra imports in conf.py of the previous themes
