### 2024-05-02 8a7bdc2
Update the notebooks examples automatically upon RTD build
