### 2024-05-02 #6418 @Bachibouzouk
> 2\. the navigation on the left sidebar (with the partial TOC) is a bit counter-intuitive, I find. I am getting lost easily without knowing where I am. If you agree, I would suggest increasing the scope of the sidebar's TOC.

@FabianHofmann - this is addressed in https://github.com/RAMP-project/RAMP/pull/137, you can see the built documentation in the PR checks, here is the direct link as well : https://rampdemand-rtd--137.org.readthedocs.build/en/137/intro.html
