### 2024-05-02 1b229a7
Copy README.rst from repo source and adapt link to contributing
As the contribution guidelines have a dedicated in the documentation we
replace the link within the readme which was pointing towards the file
CONTRIBUTING.md on github with a link which points towards the dedicated
contribution guidelines within the documentation
