+++ docs/source/intro.rst
@@ -3,5 +3,5 @@
 RAMP Introduction
 *****************
 
-.. include:: ../../README.rst
+.. include:: readme.rst
 
