### 2024-05-02 e6b9731
Add CONTRIBUTING.md content to RTD in a dedicated section
