### 2024-05-02 #6418 @Bachibouzouk
> 2\. I find it a little confusing how the [documentation site contribution guidelines](https://rampdemand.readthedocs.io/en/latest/intro.html#contributing) and file [`CONTRIBUTING.md`](https://github.com/RAMP-project/RAMP/blob/main/CONTRIBUTING.md) in the repository don't match?

@trevorb1 - this should be now fixed in https://github.com/RAMP-project/RAMP/pull/138
