### 2024-05-02 27de4f1
Link contributing guidelines within issue template
