+++ .github/ISSUE_TEMPLATE/custom.md
@@ -1,6 +1,6 @@
 ---
 name: Custom issue template
-about: Describe this issue template's purpose here.
+about: If your issue is not a bug
 title: ''
 labels: ''
 assignees: ''
@@ -8,3 +8,9 @@ assignees: ''
 ---
 
 
+Providing the following information will help RAMP developers assisting you:
+- [ ] OS: *your_operating_system*, *your_distribution*
+- [ ] RAMP version or Branch: *RAMP version* or *branch_name*, updated on *update_date*
+- [ ] If applicable: Share used input data
+- 
+<sub>*For more information on how to contribute, please check the [Contributing Guidelines](https://github.com/RAMP-project/RAMP/blob/main/CONTRIBUTING.md).*<sub>
