### 2024-05-02 bc768cc
Link contributing guidelines within issue template
