+++ .github/pull_request_template.md
@@ -1,0 +1,6 @@
+Fix #Issue
+
+**Changes proposed in this pull request**:
+- Your_changes
+
+<sub>*For more information on how to contribute, please check the [Contributing Guidelines](https://github.com/RAMP-project/RAMP/blob/main/CONTRIBUTING.md).*<sub>
