### 2024-05-02 #6418 @Bachibouzouk
> 4\. In contributions guidelines, it may be good to specify what to include in a new issue ticket and PR (or even create templates for these)? For example, for a new issue, do you want to know OS the user is running, version of RAMP they are running, ect.

@trevorb1 - this has been added in https://github.com/RAMP-project/RAMP/pull/135
