### 2024-05-06 cb2f955
Remove deprecated code usage
