+++ ramp/example/input_file_2.py
@@ -24,7 +24,7 @@ User_list.append(HH)
 HH_shower_P = pd.read_csv("ramp/example/shower_P.csv")
 
 # High-Income
-HH_shower = HH.Appliance(1, HH_shower_P, 2, 15, 0.1, 3, thermal_P_var=0.2)
+HH_shower = HH.add_appliance(1, HH_shower_P, 2, 15, 0.1, 3, thermal_P_var=0.2)
 HH_shower.windows([390, 540], [1080, 1200], 0.2)
 
 
