### 2024-05-06 458487c
Copy the appliance parameter docstrings to add_appliance docstring
