### 2024-05-06 307af07
Copy the appliance parameter docstrings to add_appliance docstring
