+++ docs/source/examples/simple_bulb/simple_bulb.rst
@@ -1,14 +1,24 @@
 Simple Appliances with multiple functioning windows
 ===================================================
 
+In this example we are going to build a load simulation of a households
+category with only access to some indoor light bulbs.
+
 .. code:: ipython3
 
     # importing functions
-    from ramp import User, UseCase, get_day_type
+    from ramp import User, UseCase
     import pandas as pd
 
 Creating a user category
 ~~~~~~~~~~~~~~~~~~~~~~~~
+
+To represent a user category, which is a collection of users with
+similar consumption characteristics such as appliance ownership and
+consumption behavior, the class **User** should be used. Each User
+instance is initially characterized by the number of users assigned to
+the user category. For example, let’s consider a household category
+comprising 10 households:
 
 .. code:: ipython3
 
@@ -20,20 +30,58 @@ Creating a user category
 Creating a simple appliance with two functioning time
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
+In the most basic use cases of RAMP, you can model appliances with
+random usage variability across different time windows throughout the
+day, as will be created in this example.
+
+To model appliances within a household category, parameters such as the
+number of appliances owned, power usage, number of usage windows, and
+total usage time need to be specified. Additionally, parameters like the
+minimum usage time after a switch-on event can enhance the realism of
+simulations. The randomness of appliance usage can be adjusted through
+parameters such as variability in the usage window or total usage time.
+
+For example, let’s consider a household category where each household
+owns 6 “Indoor Light Bulbs,” each consuming 7 Watts. These light bulbs
+are used for 120 minutes in two time windows: from 00:00 to 00:30 and
+from 19:30 to 24:00. In RAMP simulations, time resolution is measured in
+minutes, with each simulation representing a single day from 00:00
+(corresponding to minute 0) to 24:00 (corresponding to minute 1440).
+
+Assuming a minimum usage time of 10 minutes each time a light bulb is
+switched on, we can simulate this appliance by creating a new appliance
+using the **add_appliance** method of the User object:
+
 .. code:: ipython3
 
+    # add_appliance is meth
     indoor_bulb = household.add_appliance(
-        name="Indoor Light Bulb",
-        number=6,
-        power=7,
-        num_windows=2,
-        func_time=120,
-        time_fraction_random_variability=0.2,
-        func_cycle=10,
-        window_1=[1170, 1440],  # from 19:30 to 24:00
-        window_2=[0, 30],  # from 24 to 00:30
-        random_var_w=0.35,
+        name="Indoor Light Bulb", # the name of the appliance
+        number=6, # how many of this appliance each user has in this user category
+        power=7, # the power (in Watt) of each single appliance. RAMP does not deal with units of measures, you should check the consistency of the unit of measures throughout your model
+        num_windows=2, # how many usage time windows throughout the day?
+        func_time=120, # the total usage time of appliances
+        func_cycle=10, # the minimum usage time after a switch on event
+        window_1=[0, 30],  # from 24 to 00:30
+        window_2=[1170, 1440],  # from 19:30 to 24:00
+        random_var_w=0.35, # Variability of the windows in percentage
+        time_fraction_random_variability=0.2, # randomizes the total time the appliance is on (between 0 and 1)
     )
+
+.. code:: ipython3
+
+    print(household)
+
+
+.. parsed-literal::
+
+       user_name  num_users               name  number  power
+    0  Household         10  Indoor Light Bulb       6    7.0
+
+
[…]
