+++ docs/source/examples/parallel_processing/parallel_processing.rst
@@ -154,3 +154,5 @@ You can also set the ``parallel_processing`` attribute of the
 .. image:: output_10_2.png
 
 
+
+:download:`Link to the jupyter notebook file </../notebooks/parallel_processing.ipynb>`.
