+++ docs/source/examples/thermal_app/thermal_app.rst
@@ -1,20 +1,18 @@
 Thermal loads
 =============
 
-This example input file represents a single household user whose only
-load is the “shower”. The example showcases how to model thermal loads
-by: 1) using a time-varying average ``power`` attribute, pre-calculated
-as a function of the average daily groundwater temperature; and 2) using
-the ``thermal_p_var`` attribute to add further variability to the actual
-power absorbed by the appliance in each usage event, which reflects the
-randomness of user behaviour in preferring a slightly warmer or colder
-shower temperature.
+Some of the energy requirements of users are drived my technologies that
+have time variant power consumption. Hot water boilers are an example of
+technologies that can have variant powers throughout time. To show case
+how loads as such can be modelled in RAMP, we take the example of a
+household thermal load for generating shower hot water.
 
 .. code:: ipython3
 
     # importing functions
-    from ramp import User, UseCase, load_data, get_day_type
+    from ramp import User, UseCase, load_data
     import pandas as pd
+    import matplotlib.pyplot as plt
 
 Creating a user category and appliances
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -23,18 +21,16 @@ Creating a user category and appliances
 
     household = User()
 
-When the power varies across days of the year, for instance, as a
-function of the average daily groundwater temperature, the “power”
-parameter can be passed as a ``pd.DataFrame`` or ``np.array`` with a
-daily profile (365 rows of data). For this exercise, data can be loaded
-from the default examples in ramp:
+RAMP allows the user to give variant power profile for an appliance with
+a daily resolution. In this case,the power property in an appliance
+should be given as a timeseries pd.DataFrame or np.array with daily
+profile (365 rows of data). For this exercise, we will use a built-in
+power profile for thermal load example that can be loaded using
+load_data util function:
 
 .. code:: ipython3
 
     shower_power = load_data(example="shower")
-
-.. code:: ipython3
-
     # shower power distribution
     shower_power.plot()
 
@@ -48,7 +44,7 @@ from the default examples in ramp:
 
 
 
-.. image:: output_6_1.png
+.. image:: output_5_1.png
 
 
 .. code:: ipython3
@@ -56,14 +52,12 @@ from the default examples in ramp:
     shower = household.add_appliance(
         name="Shower",
         number=1,
-        power=shower_power,
-        num_windows=2,
-        func_time=15,
-        time_fraction_random_variability=0.1,
-        func_cycle=3,
-        window_1=[390, 540],
-        window_2=[1080, 1200],
-        random_var_w=0.2,
+        power=shower_power, # pass the pd.DataFrame or np.array instead of a number
+        num_windows=2, # two possibe time window for shower
+        func_time=15, # each shower takes 15 minute
+        func_cycle=3, # every
+        window_1=[390, 540], #  morning shower from 6:30 to 9:00 AM
+        window_2=[1080, 1200], # evening shower from 18:00 to 20:00
     )
 
 Generating profiles for increasing degrees of ``thermal_p_var``
@@ -80,20 +74,39 @@ Generating profiles for increasing degrees of ``thermal_p_var``
     You will simulate 365 day(s) from 2020-01-01 00:00:00 until 2020-12-31 00:00:00
 
 
+.. parsed-literal::
+
[…]
