### 2024-05-08 #6418 @Bachibouzouk
> 1. there are some examples which do not have an intro at the beginning, like the fixed flat-appliance example. These are however very valuable to understand the context.

@FabianHofmann - This should have been addressed in https://github.com/RAMP-project/RAMP/pull/124
