+++ ramp/__init__.py
@@ -13,7 +13,6 @@ Package dependencies:
     - random
 """
 
-
 from ramp._version import __version__
 from ramp.core.core import UseCase, User, Appliance
 from ramp.core.utils import yearly_pattern, get_day_type
