### 2024-05-08 bfcc1d0
Fix test setup
