### 2024-05-08 e8fc3b4
Add coveralls badge to readme
