+++ RAMP-GUI/RAMP-GUI_Modifiable/RAMP.py
@@ -1,0 +1,1629 @@
+
+"""
+Created on Thu Jul  6 19:19:52 2023
+
+@author: Bhavesh Soni
+"""
+
+
+
+
+from PyQt5 import QtCore, QtGui, QtWidgets, uic
+from PyQt5.QtWidgets import QFileDialog, QHeaderView, QTableWidgetItem, QApplication, QMainWindow, QTreeView, QPushButton, QTextEdit, QWidget
+from PyQt5.QtGui import QStandardItemModel, QStandardItem, QFont
+from PyQt5.QtCore import Qt
+
+from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
+import matplotlib.pyplot as plt
+
+from ramp import User, Appliance, UseCase, get_day_type
+import pandas as pd
+from openpyxl import load_workbook
+from openpyxl.worksheet.cell_range import CellRange
+
+import matplotlib
+import matplotlib.pyplot as plt
+
+import os
+from contextlib import redirect_stdout, redirect_stderr
+from io import StringIO
+import sys
+import shutil
+
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+
+
+# Will Show Errors and Messege in MessegeBox
+class QTextEditLogger(QTextEdit):
+    def __init__(self, parent=None):
+        super().__init__(parent)
+        self.setReadOnly(True)  # Set read-only mode
+
+        # Set size and read-only mode
+        self.setFixedSize(1850, 200)  # Adjust the size according to your needs
+        self.setReadOnly(True)
+
+        font = QFont()
+        font.setPointSize(10)
+        self.setFont(font)
+        self.setVerticalScrollBarPolicy(Qt.ScrollBarAlwaysOn)
+
+    def write(self, message):
+        self.append(message)
+        
+
+
+
+# Main Window Class
+class Ui_MainWindow(QtWidgets.QMainWindow):
+       
+    
+    
+    def __init__(self):
+        super(Ui_MainWindow, self).__init__()
+        uic.loadUi('RAMP.ui', self)
+
+        # Initialize the QTextEditLogger
+        self.log_textedit = QTextEditLogger(self.console_1)
+        sys.stdout = self.log_textedit
+        sys.stderr = self.log_textedit
+        
+        # Initialize the rest of your setup here
+        self.setupUi(self)
+
+
+
+    def setupUi(self, MainWindow):
+
+
+        self.use_case = None
+        self.selected_users = []
+
+
+        # Keep All applaince widget hidden initially
+        for i in range(1, 31):
+            getattr(self, f'widget_{i}').hide()
+
+        
+        
+        # Set default 'Whole Week' for combo box for each Applainces (RAMP Lite)
+        for i in range(1, 31):
+            getattr(self, f'wd_we_comboBox_{i}').setCurrentIndex(2)
+        
+        
+
+        
+        # Set widgets disabled at initial stage (RAMP LIte)
+        self.start_dateEdit_1.setEnabled(False)
+        self.ndays_spinBox_1.setEnabled(False)
+        self.set_title_lineEdit_1.setEnabled(False)
+        self.simulate_pushButton_1.setEnabled(False)     
+        self.scrollArea.setEnabled(False)
+        self.console_1.setSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Expanding)
+        self.console_1.setMinimumSize(200, 200) 
+     
+        # PushButtons connected to functions (RAMP Lite)
+        #
+        self.save_image_pushButton.clicked.connect(self.save_figure_lite)
+        self.generate_pushButton_1.clicked.connect(self.generate_plots_lite)
[…]
