+++ RAMP-GUI/RAMP-GUI_Stand-alone/Read_me.txt
@@ -1,16 +1,15 @@
-This project is open-source. Interested users are therefore invited to test, comment or contribute to the tool.
+RAMP-GUI Standalone python file
+Disclaimer : This project is open-source. Interested users are therefore invited to test, comment or contribute to the tool.
 
 Requirement : Python 3 and above
 Package required : rampdemand, PyQt5 and all other pre installed standard packages
 
-To modify the app according to your needs, you may use .py and .gui files. To modify or make changes in .gui file you may need Qt Designer or You may just make changes in .gui itself by opening it in python. 
-
+To modify the GUI according to your needs, you may use RAMP.py file.
 In the begining, inside .py file there are list of imported libraries or packages. All those required packages must be installed in python.
-
-If you change anything or modify anything according to your needs, you will require to bundle it together to use it as a standalone .exe file. It can also be used by running .py file without bundeling it. 
-
 
 This application was developed using Qt Designer & PyQt5.
 
 PyQt5: Riverbank Computing.
 Qt: The Qt Company Ltd.
+
+@Bhavesh Soni
