+++ RAMP-GUI/RAMP-GUI_Stand-alone/Read_me.txt
@@ -5,9 +5,9 @@ Requirement : Python 3 and above
 Package required : rampdemand, PyQt5 and all other pre installed standard packages
 
 To modify the GUI according to your needs, you may use RAMP.py file.
-In the begining, inside .py file there are list of imported libraries or packages. All those required packages must be installed in python.
+In the begining, inside .py file there are list of imported packages. All those required packages must be installed in python.
 
-This application was developed using Qt Designer & PyQt5.
+This Project was developed using Qt Designer & PyQt5.
 
 PyQt5: Riverbank Computing.
 Qt: The Qt Company Ltd.
