+++ README.rst
@@ -194,11 +194,12 @@ Other options are documented in the help of `ramp`, which you access with the ``
 
 
 If you have existing python input files, you can convert them to
-spreadsheet. To do so, go to the ``\ramp`` folder and run
+spreadsheet. Simply run
 
 .. code-block:: bash
 
-   python ramp_convert_old_input_files.py -i <path to the input file you wish to convert>
+   ramp_convert -i <path to the input file you wish to convert>
+
 
 For other examples of command lines options, such as setting date ranges, please visit `the dedicated section  <https://rampdemand.readthedocs.io/en/latest/examples/year_simulation/year_simulation.html#setting-date-range>`_ of the documentation.
 
