### 2024-05-13 #6418 @Bachibouzouk
> 3\. What is the test coverage of the project? Since there are not that many, it would be good to check that minimum coverage is realised.

@FabianHofmann - The test coverage has been realized for the first time and is stated to be 75.745% in https://github.com/RAMP-project/RAMP/pull/136
