+++ ramp/example/input_file_3.py
@@ -25,69 +25,73 @@ User_list.append(HH)
 # Create Cooking appliances
 
 HH_lunch1_soup = HH.add_appliance(
-    1, 1800, 2, 70, 0.15, 60, thermal_P_var=0.2, pref_index=1, fixed_cycle=1
+    1, 1800, 2, 70, 0.15, 60, thermal_p_var=0.2, pref_index=1, fixed_cycle=1
 )
 HH_lunch1_soup.windows([12 * 60, 15 * 60], [0, 0], 0.15)
 HH_lunch1_soup.specific_cycle_1(1800, 10, 750, 60, 0.15)
 HH_lunch1_soup.cycle_behaviour([12 * 60, 15 * 60], [0, 0])
 
 HH_lunch2_rice = HH.add_appliance(
-    1, 1800, 2, 25, 0.15, 20, thermal_P_var=0.2, pref_index=2, fixed_cycle=1
+    1, 1800, 2, 25, 0.15, 20, thermal_p_var=0.2, pref_index=2, fixed_cycle=1
 )
 HH_lunch2_rice.windows([12 * 60, 15 * 60], [0, 0], 0.15)
 HH_lunch2_rice.specific_cycle_1(1800, 10, 750, 15, 0.15)
 HH_lunch2_rice.cycle_behaviour([12 * 60, 15 * 60], [0, 0])
 
-HH_lunch2_egg = HH.add_appliance(1, 1200, 2, 3, 0.2, 3, thermal_P_var=0.2, pref_index=2)
+HH_lunch2_egg = HH.add_appliance(1, 1200, 2, 3, 0.2, 3, thermal_p_var=0.2, pref_index=2)
 HH_lunch2_egg.windows([12 * 60, 15 * 60], [0, 0], 0.15)
 
 HH_lunch2_platano = HH.add_appliance(
-    1, 1800, 2, 10, 0.15, 5, thermal_P_var=0.2, pref_index=2, fixed_cycle=1
+    1, 1800, 2, 10, 0.15, 5, thermal_p_var=0.2, pref_index=2, fixed_cycle=1
 )
 HH_lunch2_platano.windows([12 * 60, 15 * 60], [0, 0], 0.15)
 HH_lunch2_platano.specific_cycle_1(1800, 5, 1200, 5, 0.15)
 HH_lunch2_platano.cycle_behaviour([12 * 60, 15 * 60], [0, 0])
 
-HH_lunch2_meat = HH.add_appliance(1, 1200, 2, 7, 0.15, 3, thermal_P_var=0.2, pref_index=2)
+HH_lunch2_meat = HH.add_appliance(
+    1, 1200, 2, 7, 0.15, 3, thermal_p_var=0.2, pref_index=2
+)
 HH_lunch2_meat.windows([12 * 60, 15 * 60], [0, 0], 0.15)
 
 HH_lunch3_beansnrice = HH.add_appliance(
-    1, 1800, 2, 45, 0.2, 30, thermal_P_var=0.2, pref_index=3, fixed_cycle=1
+    1, 1800, 2, 45, 0.2, 30, thermal_p_var=0.2, pref_index=3, fixed_cycle=1
 )
 HH_lunch3_beansnrice.windows([12 * 60, 15 * 60], [0, 0], 0.15)
 HH_lunch3_beansnrice.specific_cycle_1(1800, 10, 750, 35, 0.2)
 HH_lunch3_beansnrice.cycle_behaviour([12 * 60, 15 * 60], [0, 0])
 
-HH_lunch3_meat = HH.add_appliance(1, 1200, 2, 10, 0.2, 5, thermal_P_var=0.2, pref_index=3)
[…]
