### 2024-05-13 0fefa91
Explicit the arguments in methods the example files
Do this only for the first occurence as there are many appliances and
users in examples 1 and 3
