### 2024-05-14 652f6df
Freeze black version for CI and add jupyter linting
