+++ .github/workflows/pytest.yml
@@ -26,7 +26,7 @@ jobs:
     - name: Install black
       run: |
         python -m pip install --upgrade pip
-        pip install black
+        pip install black[jupyter]==24.4.2
     - name: Lint with black
       run: |
         black . --check
