+++ ramp/example/input_file_1.py
@@ -46,7 +46,16 @@ User_list.append(Church)
 # Create new appliances
 
 # Church
-Ch_indoor_bulb = Church.add_appliance(number=10, power=26, num_windows=1, func_time=210, time_fraction_random_variability=0.2, func_cycle=60, fixed="yes", flat="yes")
+Ch_indoor_bulb = Church.add_appliance(
+    number=10,
+    power=26,
+    num_windows=1,
+    func_time=210,
+    time_fraction_random_variability=0.2,
+    func_cycle=60,
+    fixed="yes",
+    flat="yes",
+)
 Ch_indoor_bulb.windows(window_1=[1200, 1440], window_2=[0, 0], random_var_w=0.1)
 
 Ch_outdoor_bulb = Church.add_appliance(7, 26, 1, 150, 0.2, 60, "yes", flat="yes")
@@ -59,7 +68,9 @@ Ch_speaker.windows([1200, 1350], [0, 0], 0.1)
 Pub_lights = Public_lighting.add_appliance(12, 40, 2, 310, 0.1, 300, "yes", flat="yes")
 Pub_lights.windows([0, 336], [1110, 1440], 0.2)
 
-Pub_lights_2 = Public_lighting.add_appliance(25, 150, 2, 310, 0.1, 300, "yes", flat="yes")
+Pub_lights_2 = Public_lighting.add_appliance(
+    25, 150, 2, 310, 0.1, 300, "yes", flat="yes"
+)
 Pub_lights_2.windows([0, 336], [1110, 1440], 0.2)
 
 
