+++ ramp/example/input_file_3.py
@@ -25,7 +25,15 @@ User_list.append(HH)
 # Create Cooking appliances
 
 HH_lunch1_soup = HH.add_appliance(
-    number=1, power=1800, num_windows=2, func_time=70, time_fraction_random_variability=0.15, func_cycle=60, thermal_p_var=0.2, pref_index=1, fixed_cycle=1
+    number=1,
+    power=1800,
+    num_windows=2,
+    func_time=70,
+    time_fraction_random_variability=0.15,
+    func_cycle=60,
+    thermal_p_var=0.2,
+    pref_index=1,
+    fixed_cycle=1,
 )
 HH_lunch1_soup.windows(window_1=[12 * 60, 15 * 60], window_2=[0, 0], random_var_w=0.15)
 HH_lunch1_soup.specific_cycle_1(p_11=1800, t_11=10, p_12=750, t_12=60, r_c1=0.15)
