### 2024-05-14 #6418 @Bachibouzouk
> 7\. The examples are great at walking the reader through the many different functions of RAMP! However, I think in the first example it would be beneficial to explicitly write out what all the different arguments in the `User(...)` and `add_appliance(...)` calls are doing. This can maybe be done through a top level description like in some of your other examples (as also suggested by @FabianHofmann). At first, especially for the `add_appliance(...)` method, it wasn't necessarily clear to me what the different arguments did. While I eventually did find the info I was looking for in the `Appliance.__init__` API ref, this was not the first place I intuitively thought to look for the information (which was the `add_appliance(...)` method API ref).

@trevorb1 - https://github.com/RAMP-project/RAMP/pull/142 aims to address your comments, I added a docstring to `add_appliance()` and I provided the kwargs in the example files (the ones under `ramp/examples`)
