### 2024-05-14 #6418 @trevorb1
Hi @Bachibouzouk; apologies in the delay for getting back to you!

> 2. I find it a little confusing how the [documentation site contribution guidelines](https://rampdemand.readthedocs.io/en/latest/intro.html#contributing) and file [CONTRIBUTING.md](https://github.com/RAMP-project/RAMP/blob/main/CONTRIBUTING.md) in the repository don't match?
@trevorb1 - this should be now fixed in https://github.com/RAMP-project/RAMP/pull/138

Thank you for unifying the contribution guidelines; looks great to me!

> 4. In contributions guidelines, it may be good to specify what to include in a new issue ticket and PR (or even create templates for these)? For example, for a new issue, do you want to know OS the user is running, version of RAMP they are running, ect.
@trevorb1 - this has been added in https://github.com/RAMP-project/RAMP/pull/135

Thank you for adding the Issue and PR templates; they look great to me!

> 1. In the Quick Start section, appliances are added through the method User.Appliance(...). Later on in the examples section, appliances are added through the method User.add_appliance(...). Based on the API reference, the Appliance(...) method should only be used for when working with legacy code? If users are encouraged to use the add_appliance(...) method, updating docs to reflect this would be good! (Same issue also appears in ["Appliances with multiple cycles"](https://rampdemand.readthedocs.io/en/latest/examples/multi_cycle/multi_cycle.html) example)

> 7. The examples are great at walking the reader through the many different functions of RAMP! However, I think in the first example it would be beneficial to explicitly write out what all the different arguments in the User(...) and add_appliance(...) calls are doing. This can maybe be done through a top level description like in some of your other examples (as also suggested by @FabianHofmann). At first, especially for the add_appliance(...) method, it wasn't necessarily clear to me what the different arguments did. While I eventually did find the info I was looking for in the Appliance.__init__ API ref, this was not the first place I intuitively thought to look for the information (which was the add_appliance(...) method API ref).

> @trevorb1 - https://github.com/RAMP-project/RAMP/pull/142 aims to address your comments, I added a docstring to add_appliance() and I provided the kwargs in the example files (the ones under ramp/examples)

Thank you for adding the API ref under `add_appliance()` and updating `Appliacnce` to `add_appliance` in the examples. Looks great to me!

From my perspective, documentation points 1,2,4 and 7 have now all been fully addressed. Thanks!
