### 2024-05-15 eb71971
Fix pandas deprecation warnings and update doc
