### 2024-05-15 23e793f
Fix pandas deprecation warnings and update doc
