### 2024-05-15 b27d0ff
Improve docstring
