### 2024-05-15 c8d6854
Improve README section about conversion from py to xlsx
