### 2024-05-15 #6418 @Bachibouzouk
> 2\. JOSS asks for reproducibility of the figure in the paper (Data Sharing & Reproducibility). Is the code online somewhere? I could not see it in the `joss_paper`  branch of the repo.

> 1. Same question as @FabianHofmann; is there code where I can reproduce Figure 1?

@FabianHofmann , @trevorb1 - this is addressed in https://github.com/RAMP-project/RAMP/pull/140 we introduced a seed for the random number generator in https://github.com/RAMP-project/RAMP/pull/128 such that we can make sure figures are 100% reproducable
