### 2024-05-15 6e90793
Update CONTRIBUTING.md with developper setup
Pre-commit already getting rid of trailing spaces
