### 2024-05-15 #6418 @Bachibouzouk
> 3\. It may be worth noting in the contributing guidelines that developers should follow Black formatting, and contributions will be checked for it with your actions (although, I do see you have the Black badge on the README.md). Alternatively, having a pre-commit file, or specifying in the contributing guidelines how to install Black in VSCode (or similar) to autolint contributions would be good!

@trevorb1 - this is now addressed in https://github.com/RAMP-project/RAMP/pull/143
