### 2024-05-17 #6418 @FLomb
@trevorb1 @FabianHofmann, I'd appreciate a quick clarification on one point: when you ask for a comparison with "similar software/tools", do you mean other _open_ tools or even non-open ones? Thanks.

> At the bottom of the introduction page, there is a note that says "This project is under active development!". Can you please clarify what the active development means? Just want to ensure this does not mean we can't trust the results. (Sorry, I know this one is a little pedantic!)

@trevorb1, thanks for pointing out that "active development" may not be self-explanatory for everyone. What we mean by it is that the project is actively maintained (e.g., if anyone opens an issue, this is taken into consideration and dealt with based on priorities) and continuously developed with new, enhanced features. In other words, while the core functionalities are stable and reliable, new features and improvements pop up regularly thanks to the contribution of the core development team as well as occasional external contributors. We saw the term "active development" being regularly used in various other repositories with this meaning, as opposed to projects "not being actively developed" anymore. We could perhaps make it more explicit with the following rewording:

"_This project is actively maintained and developed. This means that while we provide stable and reliable software releases, we keep developing new features and improvements for upcoming, upgraded versions of the software._"
