### 2024-05-21 #6418 @Bachibouzouk
@AdamRJensen -  is it still possible to amend the list of authors of the JOSS paper? If yes there are two people we would like to add. One of them contributed to refactoring the code which lead the the version of RAMP we are submitting to JOSS and the other one did fix several bugs and opened interesting discussions in issues quite recently. Let me know if this would be ok on your side and I will provide their information.
