### 2024-05-21 #6418 @trevorb1
>this is addressed in https://github.com/RAMP-project/RAMP/pull/140 we introduced a seed for the random number generator in https://github.com/RAMP-project/RAMP/pull/128 such that we can make sure figures are 100% reproducable

Thanks, @Bachibouzouk! Adding the seed for the RNG looks great. My first point on general checks is considered complete.

> this is now addressed in https://github.com/RAMP-project/RAMP/pull/143

Thanks @Bachibouzouk for adding in formatting checks! My third point on documentation is considered complete.

> I'd appreciate a quick clarification on one point: when you ask for a comparison with "similar software/tools", do you mean other open tools or even non-open ones? Thanks.

@FLomb From my perspective, I think it would depend a little on the state of the field. Since I am not that familar with the state-of-the-art for stochastic load generators, from the paper I am looking for something that describes how RAMP fits into the current literature. If there are similar tools, but only propitiatory, then mentioning this would be helpful. If instead, there are similar tools that are open, mentioning how RAMP is different from them is needed. Im not necessarily looking for a huge comparison table with every similar tool; rather, just what other main tools exist for stochastic load generation and what literature/software gap RAMP fills.

I get the feeling you allude to it with the sentences in the paper:

>In fact, the software is designed to require only a basic understanding of the expected user activity patterns and owned appliances as inputs, to be provided in tabular (.xlsx) format. Then, it leverages stochasticity (using the random package) to make up for the lack of more detailed information and to account for the unpredictability of human behaviour.

Just expanding this to explicitly say what the software gap is. For example, if another tool similar to RAMP requires an proprietary input format that limits it use-case, this would be good to mention. Or if a different tool is limited to a certain sector or end-use, this would be good to note.

Please let me know if this answers your question, or if more clarification is needed!

> thanks for pointing out that "active development" may not be self-explanatory for everyone. What we mean by it is that the project is actively maintained (e.g., if anyone opens an issue, this is taken into consideration and dealt with based on priorities) and continuously developed with new, enhanced features. In other words, while the core functionalities are stable and reliable, new features and improvements pop up regularly thanks to the contribution of the core development team as well as occasional external contributors. We saw the term "active development" being regularly used in various other repositories with this meaning, as opposed to projects "not being actively developed" anymore. We could perhaps make it more explicit with the following rewording:

> _"This project is actively maintained and developed. This means that while we provide stable and reliable software releases, we keep developing new features and improvements for upcoming, upgraded versions of the software."_

Thank you for the clarification, @FLomb! Your point of "We saw the term 'active development' being regularly used in various other repositories with this meaning" is totally fair. I prefer your suggested reword, as I think it is clearer, however, this is a little subjective. Therefore, now that I have extra clarification, I will consider my point 6 on documentation complete. Regardless if you update the wording or not. :)
