### 2024-05-22 #6418 @Bachibouzouk
@AdamRJensen @trevorb,  I  summarized which comments from @trevorb1 are already addressed

## General checks

> Thanks, @Bachibouzouk! Adding the seed for the RNG looks great. My first point on general checks is considered complete.

> Thank you so much for the detailed description, @FLomb! This fully answers my question

All comments have been addressed

## Functionality

> @Bachibouzouk; thanks for the updates! From my perspective, points 2 and 3 under my functionality comments have now been addressed.

With https://github.com/RAMP-project/RAMP/pull/126#issuecomment-2046507167 all comments have been addressed

## Documentation

> From my perspective, documentation points 1,2,4 and 7 have now all been fully addressed. Thanks!

> Thanks @Bachibouzouk for adding in formatting checks! My third point on documentation is considered complete.

> Therefore, now that I have extra clarification, I will consider my point 6 on documentation complete. Regardless if you update the wording or not. :)

With https://github.com/RAMP-project/RAMP/pull/129#issuecomment-2050345429 all comments have been addressed

## Software paper

Points 1 and 2 still need to be addressed, work in progress here :)
