### 2024-05-22 c3994bc
More literature and examples
