+++ joss_paper/paper.md
@@ -64,7 +64,11 @@ The urgency of the energy transition is leading to a rapid evolution of energy s
 
 # Statement of need
 
-RAMP is an open-source, Python-based software suite that enables the stochastic simulation of any user-driven energy demand time series based on few simple inputs. In fact, the software is designed to require only a basic understanding of the expected user activity patterns and owned appliances as inputs, to be provided in tabular (`.xlsx`) format. Then, it leverages stochasticity (using the `random` package) to make up for the lack of more detailed information and to account for the unpredictability of human behaviour (see Figure \ref{fig:example}). This way, RAMP allows generating and visualising synthetic data wherever metered data does not exist, such as when designing systems in remote areas [@lombardi_generating:2019] or when looking at future electric-vehicle fleets [@mangipinto_impact:2022]. Moreover, it features several degrees of customisations, allowing users to explicitly simulate radically different but equally plausible behaviour scenarios as a key ingredient to robust system design.
+RAMP is an open-source, Python-based software suite that enables the stochastic simulation of any user-driven energy demand time series based on few simple inputs. In fact, the software is designed to require only a basic understanding of the expected user activity patterns and owned appliances as inputs, to be provided in tabular (`.xlsx`) format. For instance, a minimal definition of a user type (e.g., a certain category of households) requires only information about which energy-consuming devices they own, when, on a typical day, they tend to use them, and for how long in total. Then, the software leverages stochasticity (using the `random` package) to make up for the lack of more detailed information and to account for the unpredictability of human behaviour (see Figure \ref{fig:example}). This way, RAMP allows generating and visualising synthetic data wherever detailed metered data does not exist, such as when designing systems in remote areas [@lombardi_generating:2019] or when looking at future electric-vehicle fleets [@mangipinto_impact:2022]. 
+
+This reliance on simple inputs distinguishes RAMP from comparable tools. For instance, other popular open-source demand simulation tools, such as CREST [crest_model:2015] and demod [barsanti_demod:2021], are based on extensive and context-specific input datasets from the UK and Germany, respectively, which populate the occupancy model at the core of their approach. Similarly, the Load Profile Generator model [pflugradt_loadprofilegenerator:2022], also openly available, implements a desire-driven behavioural simulation grounded on a psychological model based on German household data. Due to their data-driven nature, these models are not readily applicable to other similar contexts, such as different European countries, and are arguably inapplicable to cases, such as remote areas, with unique characteristics. Moreover, the heavy reliance on historical data serves poorly the need to accommodate the modelling of future, entirely new behaviours or devices. RAMP trades off the capability of providing detailed information about users' occupancy and activity, typical of the above data-driven models, for substantially greater modelling flexibility and context adaptability. Previous attempts at generating demand time series based on a similar approach as RAMP exist [mandelli_lpg:2016], but they are neither open-source nor capable of simulating non-electric or more sophisticated energy uses.
+
+In fact, RAMP features several degrees of customisation, enabling modellers to switch on or off features tailored to the needs of specific energy uses. For example, to represent the highly variable load profile of cooking appliances, which changes completely depending on which meal is cooked, the software allows defining many possible meal types with an associated cooking cycle; then, it leverages stochasticity to diversity which meals are cooked when. For heat-related energy uses that may be influenced by weather parameters, such as outdoor or groundwater temperature, it is possible to provide a time series of such parameters and define how they affect the default energy consumption. These and many other customisation options allow users to explicitly simulate radically different but equally plausible behaviour scenarios, including behaviours that may happen in the future and that have no relationship with past data, as a key ingredient to robust system design.
 
[…]
