### 2024-05-27 #6418 @Bachibouzouk
@AdamRJensen I will tag and list the two additional co-authors below:

@dhungelgd did a large chunk of the refactoring of the ramp v0.3.1 code, he opened a few issues and tested that the refactor code was providing the same behavior as the original code. He worked at the time on the Reiner Lemoine Institut organisation's fork : https://github.com/rl-institut/RAMP as a Master student.

@JW-Kraft did use RAMP to simulate productive use and submitted pull request to address problem that were preventing the use of RAMP for productive usecase. He also documented some issues, fixed some himself and worked on the RAMP documentation.
