### 2024-05-27 #6418 @Bachibouzouk
> @Bachibouzouk could you link to list of their PRs/issues to RAMP. And an explanation as to why they weren't included from the beginning would be nice

@dhungelgd wasn't included in the author list because at the time he worked on RAMP (2022) he was a master student at RLI and I was the person who then created the PRs towards the official RAMP repo. He left RLI before being able to be present at many RAMP-dev meetings and was therefore not known by the other RAMP developpers. When we submitted the paper in November 2023 @FLomb made the list of authors and it did not struck me at this time that @dhungelgd was not listed. It is only then I consulted recently the list of contributors (https://github.com/RAMP-project/RAMP/graphs/contributors) that I realized we did not included @dhungelgd in the authors list.

https://github.com/RAMP-project/RAMP/pull/36/commits

https://github.com/rl-institut/RAMP/issues/15
https://github.com/rl-institut/RAMP/issues/12
https://github.com/rl-institut/RAMP/issues/10

@JW-Kraft wasn't included in the author list because when we submitted he hadn't yet contributed
https://github.com/RAMP-project/RAMP/pull/132
https://github.com/RAMP-project/RAMP/pull/128
https://github.com/RAMP-project/RAMP/pull/116
https://github.com/RAMP-project/RAMP/pull/111

https://github.com/RAMP-project/RAMP/issues/127
https://github.com/RAMP-project/RAMP/issues/80
https://github.com/RAMP-project/RAMP/issues/78
https://github.com/RAMP-project/RAMP/issues/79
