### 2024-05-30 457df46
Fix formating of certain citations
And fix trailing spaces (pre-commit hooks)
