+++ joss_paper/paper.md
@@ -12,7 +12,7 @@ authors:
     corresponding: true # (This is how to denote the corresponding author)
     affiliation: 1
   - name: Pierre-François Duc
-    orcid: 
+    orcid:
     affiliation: 2
   - name: Mohammad Amin Tahavori
     orcid: 0000-0002-7753-0523
@@ -22,27 +22,27 @@ authors:
     affiliation: "4,7"
   - name: Sarah Eckhoff
     orcid: 0000-0002-6168-4835
-    affiliation: 5  
+    affiliation: 5
   - name: Maria C.G. Hart
     orcid: 0000-0002-1031-9782
-    affiliation: 5  
+    affiliation: 5
   - name: Francesco Sanvito
     orcid: 0000-0002-9152-9684
     affiliation: 1
   - name: Gregory Ireland
-    orcid: 
+    orcid:
     affiliation: "2,6"
   - name: Sergio Balderrama
-    orcid: 
+    orcid:
     affiliation: 7
   - name: Johann Kraft
-    orcid: 
+    orcid:
     affiliation: 2
   - name: Gokarna Dhungel
-    orcid: 
+    orcid:
     affiliation: 2
   - name: Sylvain Quoilin
-    orcid: 
+    orcid:
     affiliation: 4
 affiliations:
  - name: TU Delft, Faculty of Technology, Policy and Management, Delft, The Netherlands
@@ -70,9 +70,9 @@ The urgency of the energy transition is leading to a rapid evolution of energy s
 
 # Statement of need
 
-RAMP is an open-source, Python-based software suite that enables the stochastic simulation of any user-driven energy demand time series based on few simple inputs. In fact, the software is designed to require only a basic understanding of the expected user activity patterns and owned appliances as inputs, to be provided in tabular (`.xlsx`) format. For instance, a minimal definition of a user type (e.g., a certain category of households) requires only information about which energy-consuming devices they own, when, on a typical day, they tend to use them, and for how long in total. Then, the software leverages stochasticity (using the `random` package) to make up for the lack of more detailed information and to account for the unpredictability of human behaviour (see Figure \ref{fig:example}). This way, RAMP allows generating and visualising synthetic data wherever detailed metered data does not exist, such as when designing systems in remote areas [@lombardi_generating:2019] or when looking at future electric-vehicle fleets [@mangipinto_impact:2022]. 
+RAMP is an open-source, Python-based software suite that enables the stochastic simulation of any user-driven energy demand time series based on few simple inputs. In fact, the software is designed to require only a basic understanding of the expected user activity patterns and owned appliances as inputs, to be provided in tabular (`.xlsx`) format. For instance, a minimal definition of a user type (e.g., a certain category of households) requires only information about which energy-consuming devices they own, when, on a typical day, they tend to use them, and for how long in total. Then, the software leverages stochasticity (using the `random` package) to make up for the lack of more detailed information and to account for the unpredictability of human behaviour (see Figure \ref{fig:example}). This way, RAMP allows generating and visualising synthetic data wherever detailed metered data does not exist, such as when designing systems in remote areas [@lombardi_generating:2019] or when looking at future electric-vehicle fleets [@mangipinto_impact:2022].
 
[…]
