### 2024-05-30 #6418 @Bachibouzouk
> Points 1 and 2 still need to be addressed, work in progress here :)

Those points have been addressed now in https://github.com/RAMP-project/RAMP/pull/145

@AdamRJensen, we now addressed all comments from @FabianHofmann and @trevorb in the branch `joss-paper`  :)

I guess we now need to wait that they approve our changes before we proceed further, right?
