### 2024-06-02 85575fc
Editorial suggested changes to JOSS paper
