+++ joss_paper/paper.md
@@ -68,17 +68,17 @@ bibliography: paper.bib
 
 # Summary
 
-The urgency of the energy transition is leading to a rapid evolution of energy system design worldwide. In areas with widespread energy infrastructure, existing electricity, heat and mobility networks are being re-designed for carbon neutrality and are increasingly interconnected. In areas where energy infrastructure is limited, instead, networks and systems are being rapidly expanded to ensure access to energy for all. And yet, re-designing and expanding energy systems in these directions requires information on future user behaviour and associated energy demand, which is often unavailable. In fact, historical data are either entirely missing or poorly representative of future behaviour within transitioning systems. This results in the reliance on inadequate demand data, which affect system design and its resilience to rapid behaviour evolution.
+The urgency of the energy transition is leading to a rapid evolution of energy system design worldwide. In areas with widespread energy infrastructure, existing electricity, heat and mobility networks are being re-designed for carbon neutrality and are increasingly interconnected. In areas where energy infrastructure is limited, instead, networks and systems are being rapidly expanded to ensure access to energy for all. In both cases, re-designing and expanding energy systems in these directions requires information on future user behaviour and associated energy demand, yet this type of data is often unavailable. In fact, historical data are often either entirely missing or poorly representative of future behaviour within transitioning systems. This results in reliance on inadequate demand data, which affects system design and its resilience to rapid behaviour evolution.
 
 # Statement of need
 
-RAMP is an open-source, Python-based software suite that enables the stochastic simulation of any user-driven energy demand time series based on few simple inputs. In fact, the software is designed to require only a basic understanding of the expected user activity patterns and owned appliances as inputs, to be provided in tabular (`.xlsx`) format. For instance, a minimal definition of a user type (e.g., a certain category of households) requires only information about which energy-consuming devices they own, when, on a typical day, they tend to use them, and for how long in total. Then, the software leverages stochasticity (using the `random` package) to make up for the lack of more detailed information and to account for the unpredictability of human behaviour (see Figure \ref{fig:example}). This way, RAMP allows generating and visualising synthetic data wherever detailed metered data does not exist, such as when designing systems in remote areas [@lombardi_generating:2019] or when looking at future electric-vehicle fleets [@mangipinto_impact:2022].
+RAMP is an open-source, Python-based software suite that enables the stochastic simulation of any user-driven energy demand time series based on few simple inputs. In fact, the software is designed to require only a basic understanding of the expected user activity patterns and owned appliances as inputs, which are provided in tabular format (`.xlsx`). For instance, a minimal definition of a user type (e.g., a certain category of households) requires only information about which energy-consuming devices they own, when, on a typical day, they tend to use them, and for how long in total. Then, the software leverages stochasticity (using the `random` package) to make up for the lack of more detailed information and to account for the unpredictability of human behaviour (see Figure \ref{fig:example}). This way, RAMP allows generating and visualising synthetic data wherever detailed metered data does not exist, such as when designing systems in remote areas [@lombardi_generating:2019] or when looking at future electric-vehicle fleets [@mangipinto_impact:2022].
 
[…]
