### 2024-06-02 #6418 @AdamRJensen
A few editorial comments and questions:

- [ ] Is RAMP an acronym?
- [ ] Suggestions in https://github.com/RAMP-project/RAMP/pull/148
