### 2024-06-03 #6418 @FLomb
> A few editorial comments and questions:
>
>     * [ ]  Is RAMP an acronym?
>
>     * [ ]  Suggestions in [Editorial suggested changes to JOSS paper RAMP-project/RAMP#148](https://github.com/RAMP-project/RAMP/pull/148)

Thanks for the suggestions, @AdamRJensen. They look good, so I will approve the PR shortly.

Regarding whether RAMP is an acronym, it used to be, but it's not anymore. When we first developed it, the name was a (loose) acronym for 'Remote-Areas Multi-energy load Profiles'. However, since the range of applications has, in the meantime, gone way beyond only remote areas, we stopped referring to the original acronym in our material (website, docs, publications) and just kept the name. In fact, the name RAMP, per se, makes intuitive sense in the context of energy system load profile simulation, as 'ramp up' and 'ramp down' are commonly used terms to refer to sudden changes in the load.
