### 2024-06-03 #6418 @Bachibouzouk
@AdamRJensen - your suggestions are now merged into `joss-paper`
