### 2024-06-03 #6418 @Bachibouzouk
@AdamRJensen

We facing a (small) dilemma. There have been a few pull requests merged into `development` in the meantime. We have now two options:

1) make the release from `joss-paper` branch and then make another release later on with the content of those pull requests

2) merge `joss-paper` into development and make the release from there

The 4 PRs do not represent a lot of code changes and most of them even add more tests.
https://github.com/RAMP-project/RAMP/pull/116/files (a change in 2 lines of code)
https://github.com/RAMP-project/RAMP/pull/117/files (only changes to workflow files to publish to pypi)
https://github.com/RAMP-project/RAMP/pull/111/files (in this one the code is adapted to add the option to set an appliance in productive use mode)
https://github.com/RAMP-project/RAMP/pull/132/files (a change in one line of code)

In our opinion those are mainly bugfixes and therefore the joss publication could only benefit from those. Although it might raise the problem on your side that the reviewers would need to review those extra changes and it might be a time burden. We thought we still ask you if you are ok that we go for choice 2). If not we just implement choice 1) :)
