### 2024-06-03 #6418 @AdamRJensen
I think it's perfectly fine to include these recent PRs in the release that is reference in the JOSS publication
