### 2024-06-03 c56b550
Fix failing test for appliance with 0 func_time
