### 2024-06-04 2e2aa33
Update README.rst to clarify 'active development'
This PR addresses a comment by a JOSS reviewer that the statement about 'active development' at the bottom of the README may lead to misunderstandings.

The statement is now more explicit and clear.
