### 2024-06-04 0ec1e67
Update changelog
With some pre-commit linting fixes on previous releases
