### 2024-06-04 b31aaf8
Fix README formatting
So that twine check dist/* does not output errors.

Pre-commit removed trailing spaces in the file as well
