### 2024-06-05 0922874
fix path issues in test_run.py
