### 2024-06-05 522b436
Set a user's appliance name by default if empty
