### 2024-06-05 ff026c9
Add appliances name to example files
