+++ ramp/example/input_file_1.py
@@ -55,45 +55,50 @@ Ch_indoor_bulb = Church.add_appliance(
     func_cycle=60,
     fixed="yes",
     flat="yes",
+    name="indoor_bulb",
 )
 Ch_indoor_bulb.windows(window_1=[1200, 1440], window_2=[0, 0], random_var_w=0.1)
 
-Ch_outdoor_bulb = Church.add_appliance(7, 26, 1, 150, 0.2, 60, "yes", flat="yes")
+Ch_outdoor_bulb = Church.add_appliance(
+    7, 26, 1, 150, 0.2, 60, "yes", flat="yes", name="outdoor_bulb"
+)
 Ch_outdoor_bulb.windows([1200, 1440], [0, 0], 0.1)
 
-Ch_speaker = Church.add_appliance(1, 100, 1, 150, 0.2, 60)
+Ch_speaker = Church.add_appliance(1, 100, 1, 150, 0.2, 60, name="speaker")
 Ch_speaker.windows([1200, 1350], [0, 0], 0.1)
 
 # Public lighting
-Pub_lights = Public_lighting.add_appliance(12, 40, 2, 310, 0.1, 300, "yes", flat="yes")
+Pub_lights = Public_lighting.add_appliance(
+    12, 40, 2, 310, 0.1, 300, "yes", flat="yes", name="lights"
+)
 Pub_lights.windows([0, 336], [1110, 1440], 0.2)
 
 Pub_lights_2 = Public_lighting.add_appliance(
-    25, 150, 2, 310, 0.1, 300, "yes", flat="yes"
+    25, 150, 2, 310, 0.1, 300, "yes", flat="yes", name="lights2"
 )
 Pub_lights_2.windows([0, 336], [1110, 1440], 0.2)
 
 
 # High-Income
-HI_indoor_bulb = HI.add_appliance(6, 7, 2, 120, 0.2, 10)
+HI_indoor_bulb = HI.add_appliance(6, 7, 2, 120, 0.2, 10, name="indoor_bulb")
 HI_indoor_bulb.windows([1170, 1440], [0, 30], 0.35)
 
-HI_outdoor_bulb = HI.add_appliance(2, 13, 2, 600, 0.2, 10)
+HI_outdoor_bulb = HI.add_appliance(2, 13, 2, 600, 0.2, 10, name="outdoor_bulb")
 HI_outdoor_bulb.windows([0, 330], [1170, 1440], 0.35)
 
-HI_TV = HI.add_appliance(2, 60, 3, 180, 0.1, 5)
+HI_TV = HI.add_appliance(2, 60, 3, 180, 0.1, 5, name="TV")
 HI_TV.windows([720, 900], [1170, 1440], 0.35, [0, 60])
 
-HI_DVD = HI.add_appliance(1, 8, 3, 60, 0.1, 5)
+HI_DVD = HI.add_appliance(1, 8, 3, 60, 0.1, 5, name="DVD")
 HI_DVD.windows([720, 900], [1170, 1440], 0.35, [0, 60])
 
-HI_Antenna = HI.add_appliance(1, 8, 3, 120, 0.1, 5)
+HI_Antenna = HI.add_appliance(1, 8, 3, 120, 0.1, 5, name="antenna")
 HI_Antenna.windows([720, 900], [1170, 1440], 0.35, [0, 60])
 
-HI_Phone_charger = HI.add_appliance(5, 2, 2, 300, 0.2, 5)
+HI_Phone_charger = HI.add_appliance(5, 2, 2, 300, 0.2, 5, name="phone_charger")
[…]
