### 2024-06-05 14f1d61
Reselect a switch on even if falling outside duty cycle windows
