### 2024-06-05 510c164
Extend duty_cycle window to upper bound
