### 2024-06-05 fb91034
Change duty_cycle selection method
Previously the mean of the index range was computed and see if it was
falling within a duty cycle. This method compare the boundaries of the
range with the one of each duty cycle to know if there is an
intersection. If there is one, the duty cycle is chosen.

Note: this method does not take into account the case where the indexes
range spans accross more than one duty cycle.
