### 2024-06-06 75cb11d
Minor docstring improvements
