### 2024-06-06 f25fbf4
Remove use of matplotlib join method in test_use
