### 2024-06-06 37c9d21
Remove contributing.md from docs/source
