### 2024-06-07 #6418 @Bachibouzouk
> @Bachibouzouk, @FLomb Just wanted to make sure that you known that the next step is for you to make a new release of the package and upload to Zenodo (maybe you've already done that), and then write back here with the version number and doi.

@AdamRJensen - we are on it since monday, but we have encountered a few release bugs (new version of matplotlib which doesn't contain a function we were using, tests not running locally, etc..) which delayed the process. We should be able to release today though :)
