### 2024-06-08 #6418 @FLomb
@AdamRJensen, we have completed the new release of the package and its upload to Zenodo. The version is 'v0.5.2' and the Zenodo DOI is the following: [10.5281/zenodo.11526597](https://doi.org/10.5281/zenodo.11526597).
