### 2024-06-09 #6418 @AdamRJensen
> MISSING DOIs
>
> - No DOI given, and none found for title: RAMP: stochastic multi-energy demand profiles
> - 10.24251/hicss.2023.097 may be a valid DOI for title: Sustainable Energy System Planning in Developing C...
>
> INVALID DOIs
>  https://doi.org/10.1016/j.esd.2016.01.005 is INVALID because of 'https://doi.org/' prefix

@FLomb could you fix these? Also, I think instead of citing the github repo then just put the link in a parentheses in the text instead of a citation.
