### 2024-06-10 #6418 @FLomb
>
> @FLomb could you fix these? Also, I think instead of citing the github repo then just put the link in a parentheses.

@AdamRJensen I'm sorry about the missing/invalid DOIs. We had checked them, but these issues must have slipped through. How to best fix them? Should we make a new minor release with the corrected paper?
