### 2024-06-10 #6418 @AdamRJensen
> @AdamRJensen I'm sorry about the missing/invalid DOIs. We had checked them, but these issues must have slipped through. How to best fix them? Should we make a new minor release with the corrected paper?
I don't see any need to make a new release, but if you could just make the changes to the branch with the paper. The version and doi are only for the code part and not the paper part as far as I understand.
