### 2024-06-11 bacc848
Update paper.md
Changed GitHub reference from .bib to url
