### 2024-06-11 2ebce2c
Update paper.bib
Fixed DOIs and removed reference to GitHub repo
