+++ joss_paper/paper.bib
@@ -193,19 +193,10 @@ Publisher: Multidisciplinary Digital Publishing Institute},
 }
 
 
-@misc{ramp_github,
-	title = {{RAMP: stochastic multi-energy demand profiles}},
-	url = {https://github.com/RAMP-project},
-	year = {2019},
-    publisher = {GitHub},
-    journal = {GitHub repository},
-}
-
-
 @article{hart_gui:2023,
 	title = {Sustainable {Energy} {System} {Planning} in {Developing} {Countries}: {Facilitating} {Load} {Profile} {Generation} in {Energy} {System} {Simulations}},
 	isbn = {978-0-9981331-6-4},
-	url = {https://hdl.handle.net/10125/102726},
+	doi = {10.24251/hicss.2023.097},
 	author = {Hart, Maria C. G. and Eckhoff, Sarah and Breitner, Michael H.},
 	year = {2023},
     journal = {Hawaii International Conference on System Sciences}
@@ -214,7 +205,7 @@ Publisher: Multidisciplinary Digital Publishing Institute},
 
 @misc{ramp_zenodo:2023,
 	title = {{RAMP}-project/{RAMP}: v0.5.0},
-	url = {https://doi.org/10.5281/zenodo.10275752},
+	url = {10.5281/zenodo.10275752},
 	publisher = {Zenodo},
 	author = {Lombardi, Francesco and Duc, Pierre-Francois and Tahavori, Mohammad Amin and Sanchez-Solis, Claudia and Eckhoff, Sarah and Hart, Maria C.G. and Sanvito, Francesco Davide and Ireland, Gregory and Balderrama, Sergio and Quoilin, Sylvain},
 	year = {2023},
@@ -235,7 +226,7 @@ Publisher: Multidisciplinary Digital Publishing Institute},
 	title = {Novel procedure to formulate load profiles for off-grid rural areas},
 	volume = {31},
 	issn = {0973-0826},
-	doi = {https://doi.org/10.1016/j.esd.2016.01.005},
+	doi = {10.1016/j.esd.2016.01.005},
     journal = {Energy for Sustainable Development},
 	author = {Mandelli, Stefano and Merlo, Marco and Colombo, Emanuela},
 	year = {2016},
