### 2024-06-11 #6418 @FLomb
@AdamRJensen Thanks for the clarification. I have fixed the DOI issues in the `joss-paper` branch.
