### 2024-06-11 #6418 @kyleniemeyer
Hi @FLomb, it looks like one of your coauthors (Sergio Balderrama) is missing from the Zenodo author/contributor list. Can you correct that? You can edit the Zenodo archive metadata without needing a new version/DOI.
