### 2024-06-12 #6418 @FLomb
> Hi @FLomb, it looks like one of your coauthors (Sergio Balderrama) is missing from the Zenodo author/contributor list. Can you correct that? You can edit the Zenodo archive metadata without needing a new version/DOI.

Hi @kyleniemeyer, thank you for spotting it; that was quite a mistake! Sergio Balderrama has now been added to the Zenodo author list.
